add propagation

+++ src/csv_util.h
@@ -9,22 +9,22 @@
 
 // typedef Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> MatrixXi;
 
-Eigen::MatrixXi read_csv(const std::string& path) {
+Eigen::MatrixXd read_csv(const std::string& path) {
 	std::cerr << "Importing from csv: " << path << std::endl;
 	std::ifstream indata;
 	indata.open(path);
 	std::string line;
-	std::vector<int> values;
+	std::vector<double> values;
 	uint rows = 0;
 	while (std::getline(indata, line)) {
 		std::stringstream lineStream(line);
 		std::string cell;
 		while (std::getline(lineStream, cell, ',')) {
-			values.push_back(std::stoi(cell));
+			values.push_back(std::stod(cell));
 		}
 		++rows;
 	}
-	return Eigen::Map<Eigen::MatrixXi>(values.data(), rows, values.size()/rows);
+	return Eigen::Map<Eigen::MatrixXd>(values.data(), rows, values.size()/rows);
 }
 
 template<typename M>
